refactor: move grid header color style
